Added function to data forecasts
